refactor: rename extension/queue/sql to extension/queue/mysql
Aligns queue extension naming with existing conventions
(extension/storage/mysql, extension/counter/mysql). The package name
`sql` was generic; `mysql` makes the backend explicit.
