+++ extension/queue/mysql/errors.go
@@ -1,4 +1,4 @@
-package sql
+package mysql
 
 import "fmt"
