+++ extension/queue/mysql/message_store.go
@@ -1,4 +1,4 @@
-package sql
+package mysql
 
 import (
 	"context"
