+++ extension/queue/mysql/publisher.go
@@ -1,4 +1,4 @@
-package sql
+package mysql
 
 import (
 	"context"
