+++ extension/queue/mysql/schema/queue_messages.sql
