+++ extension/queue/mysql/stores.go
@@ -1,4 +1,4 @@
-package sql
+package mysql
 
 //go:generate mockgen -source=stores.go -destination=mock_stores.go -package=sql
