+++ extension/queue/mysql/subscriber.go
@@ -1,4 +1,4 @@
-package sql
+package mysql
 
 import (
 	"context"
