+++ extension/queue/mysql/subscriber_test.go
@@ -1,4 +1,4 @@
-package sql
+package mysql
 
 import (
 	"context"
