+++ extension/queue/mysql/validation.go
@@ -1,4 +1,4 @@
-package sql
+package mysql
 
 import "fmt"
