+++ test/e2e/suite_test.go
@@ -73,7 +73,7 @@ func (s *E2EIntegrationSuite) SetupSuite() {
 	testutil.ApplySchema(t, s.log, s.db, testutil.SchemaDir("extension/counter/mysql/schema"))
 
 	// Apply schemas programmatically to queue database
-	testutil.ApplySchema(t, s.log, s.queueDB, testutil.SchemaDir("extension/queue/sql/schema"))
+	testutil.ApplySchema(t, s.log, s.queueDB, testutil.SchemaDir("extension/queue/mysql/schema"))
 
 	s.log.Logf("Schemas applied successfully")
