+++ test/integration/extension/queue/mysql/BUILD.bazel
@@ -1,17 +1,17 @@
 load("@rules_go//go:def.bzl", "go_test")
 
 go_test(
-    name = "sql_test",
+    name = "mysql_test",
     srcs = ["queue_test.go"],
     data = [
         "docker-compose.yml",
-        "//extension/queue/sql/schema",
+        "//extension/queue/mysql/schema",
     ],
     tags = ["integration"],
     deps = [
         "//entity/queue",
         "//extension/queue",
-        "//extension/queue/sql",
+        "//extension/queue/mysql",
         "//test/testutil",
         "@com_github_go_sql_driver_mysql//:mysql",
         "@com_github_stretchr_testify//assert",
